refactor: doPost doGet 방식으로 handle 처리
